requirements.txt
# ── AI & HTTP ──
httpx>=0.27
python-dotenv>=1.0

# ── Report generation ──
jinja2>=3.1
pygments>=2.17
markdown>=3.5

# ── Gmail integration ──
google-api-python-client>=2.100
google-auth>=2.25
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# ── Web server ──
flask>=3.0